% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modcmfitr.R
\docType{package}
\name{modcmfitr}
\alias{modcmfitr}
\alias{modcmfitr-package}
\title{modcmfitr: A package to fit and sample from a modified Connor-Mosimann distribution}
\description{
This package fits a modified Connor-Mosimann (mCM) distribution to quantiles of a multinomial distribution
elicited from experts using a process such as the Sheffield Elicitation Framework (SHELF, O'Hagan & Oakley).
}
\details{
More details and worked examples are provided in the help files and accompanying vignette.  Functions are also provided to fit Connor-Mosimann (CM) and Dirichlet (D) distributions, and to sample from mCM and CM
distributions.  A function to sample from the Dirichlet distribution is available in the gtools package (Warnes, Bolker & Lumley).

Copyright (C) 2017  Ed Wilson, University of Cambridge, <ed.wilson@medschl.cam.ac.uk>

October 2017
}
