% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modcmfitrFunctions.R
\name{rModCM}
\alias{rModCM}
\title{Sample from modified Connor-Mosimann distribution}
\usage{
rModCM(n, Z)
}
\arguments{
\item{n}{Number of samples}

\item{Z}{vector of (k-1) scaled beta parameters.  Where k=number of dimensions to problem. Columns = (a,b,lower limit, upper limit) of scaled beta distribution}
}
\description{
This function returns n samples from a modified CM distribution
}
\examples{
n <- 1000
Z <- matrix(data = c(1.098, 3.070, 0.66, 0.85,
                   9.021, 7.990, 0.62, 1,
                   10, 0.01, 0.14, 0.58,
                   10, 10, 0.57, 1),ncol=4,byrow=TRUE)
rModCM(n,Z)
}
\keyword{Connor-Mosimann}
\keyword{modified}
